% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/immunogen_hypothesis_tools.r
\name{generate_random_hypothesis}
\alias{generate_random_hypothesis}
\title{Generate a random hypothesis}
\usage{
generate_random_hypothesis(m, upper_lim = 1)
}
\arguments{
\item{m}{Number of molecules in ELISpot population}
}
\description{
Returns a random hypothesis, used to initiate the MH chain. Based on
}

