% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/immunogen_mh_tools_2.r
\name{get_acc_rate_chain}
\alias{get_acc_rate_chain}
\title{Acceptance Rate of an MH chain}
\usage{
get_acc_rate_chain(chain)
}
\arguments{
\item{chain}{Which chain do you want to compute the acceptance rate of?}
}
\description{
Computes the acceptance rate of a chain by comparing each state to the next one.
}

