% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/immunogen_evidence.r
\name{get_evidence_1}
\alias{get_evidence_1}
\title{Evidence function}
\usage{
get_evidence_1(eli.dat, H, mol.names, pep)
}
\arguments{
\item{eli.dat}{the dataframe containing the outcome of the ELISpot experiments}

\item{H}{hypothesis, a string of values between 0 and 1 separated by commas. There should be length(mol.names) values}

\item{mol.names}{Names of the molecules in your ELISpot subjects}

\item{pep}{Name of the peptide currently being processed}
}
\description{
Computes P(D|H)
}

