% Please edit documentation in R/immunogen_decision_rules.r
\name{get_hla_ranking}
\alias{get_hla_ranking}
\title{Find pHLA combinations that explain positive ELISpot results in the cohort.}
\usage{
get_hla_ranking(peps_for_analysis, res, molecs, eli.dat, what="Mode")
}
\arguments{
\item{peps_for_analysis}{Which peptides do you want to include}
\item{res}{The results table as generated by get_overview_df}
\item{molecs}{A vector listing the molecules that are present in the ELISpot subjects}
\item{eli.dat}{The original patient and ELISpot data}
\item{what}{The summary value used to rank the HLA. Defaults to "Mode", but can also be "Mean" or "Median"}
}
\description{
For each peptide, we rank the HLAs according to their posterior mode. 
For each subject, we calculate which of their HLAs is the highest ranked one. 
In return, for each HLA, we count in how many subjects this HLA is present and in how many subjects it is the highest ranked HLA.
This calculation is then repeated, taking into account only subjets with positive ELISpot for the peptide under consideration.
This allows the user to see which HLAs 'explains' the positive ELISpot results.
}
