% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/immunogen_decision_rules.r
\name{get_KL_vs_unif}
\alias{get_KL_vs_unif}
\title{Kullback-Leibler divergence of a sample (with respect to the uniform distribution)}
\usage{
get_KL_vs_unif(sample)
}
\arguments{
\item{sample}{Sample to compute KL divergence of (in a vector)}
}
\description{
Returns the KL divergence of a sample wrt the uniform distribution
}

