% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/immunogen_decision_rules.r
\name{get_mode}
\alias{get_mode}
\title{Estimate sample mode}
\usage{
get_mode(sample)
}
\arguments{
\item{sample}{Sample to get an estimated mode of}
}
\description{
This function is a wrapper for the mlv function from the modeest package
}

