% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/immunogen_decision_rules.r
\name{get_modes_chain}
\alias{get_modes_chain}
\title{Estimate posterior modes in the MH chain}
\usage{
get_modes_chain(chain, molecs)
}
\arguments{
\item{chain}{Metropolis-Hastings chain, as obtained by mh_chain}

\item{molecs}{vector containing all the HLA-II molecules in your ELISpot subjects}
}
\description{
Returns a dataframe with the posterior mode of the pHLA immunogenicity, for each HLA.
}

