% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/immunogen_decision_rules.r
\name{get_overview_df}
\alias{get_overview_df}
\title{Overview of posterior modes, means, medians for all pHLA combinations}
\usage{
get_overview_df(peps_for_analysis, chainDir, molecs)
}
\arguments{
\item{peps_for_analysis}{Which peptides do you want to include}

\item{chainDir}{directory that contains the MH chains}

\item{molecs}{A vector listing the molecules that are present in the ELISpot subjects}
}
\description{
Returns a dataframe with the marginal posterior mode, mean and median of the pHLA immunogenicity, for each HLA.
Mode, mean and median are all on a separate line, so each pHLA combination occurs three times
Also gives the KL divergence of the marginal posterior with respect to the uniform distribution
}

