% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/immunogen_mh_tools_2.r
\name{get_proposal}
\alias{get_proposal}
\title{Choosing a proposal state}
\usage{
get_proposal(h, unif.prop = T, radius)
}
\arguments{
\item{h}{current State of the MH chain}

\item{unif.prop}{If TRUE, proposal is chosen from [0,1]^m. If false, only looks in the overlap of [0,1]^m with a cube around h with side length 2*radius}

\item{radius}{If unif.prop is FALSE, we look for a proposal state in the overlap of [0.1]^m and [h-radius, h+radius].}
}
\description{
Given the current state of the MH chain, pick a proposal state. Also returns Q(proposal state|current state)
}

