% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/immunogen_predictive_prior.r
\name{get_shape_params}
\alias{get_shape_params}
\title{Beta shape parameters}
\usage{
get_shape_params(M, SD)
}
\arguments{
\item{M}{desired mode}

\item{SD}{desired standard deviation}
}
\description{
Given a desired mode and standard deviation, compute the shape parameters for the Beta distribution
}

