% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/immunogen_mh_tools_2.r
\name{mh_chain}
\alias{mh_chain}
\title{Metropolis-Hasting for a single peptide}
\usage{
mh_chain(eli.dat, mol.names, init_h, max_steps = 5000, pep, unif.prop = T,
  radius = 0.1, p.df = NULL)
}
\arguments{
\item{eli.dat}{ELISpot dataset}

\item{mol.names}{Names of the HLA molecules in the ELISpot dataset}

\item{init_h}{Initial state of the MH chain, can be produced randomly by generate_random_hypothesis}

\item{max_steps}{The required length of the chain. The default value, 5000, is too low but this was done for debugging reasons.}

\item{pep}{Peptide currently being processed}

\item{unif.prop}{If TRUE, choose proposal states from [0,1]^m}

\item{radius}{If unif.prop is FALSE, we look for a proposal state in a hypercube around h restricted by radius. Should be trained such that the acceptance rate of the chain is around 50%, see example.r.}

\item{p.df}{Dataframe which describes the parameters of the prior distributions for each HLA}
}
\description{
Implementation of the Metropolis-Hastings algorithm to find posterior distributions of pHLA immunogenicity for a single peptide
}

