% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/immunogen_mh_tools_2.r
\name{one_time_step}
\alias{one_time_step}
\title{One update of MH chain}
\usage{
one_time_step(h, loglik_h, eli.dat, pep, mol.names, unif.prop, radius, p.df)
}
\arguments{
\item{h}{Hypothesis which is the current state of the chain}

\item{loglik_h}{Log likelihood of h}

\item{eli.dat}{ELISpot dataset}

\item{pep}{Peptide currently being processed}

\item{mol.names}{Names of the HLA molecules in the ELISpot dataset}

\item{unif.prop}{If TRUE, choose proposal state from [0,1]^m}

\item{radius}{If unif.prop is FALSE, we look for a proposal state in a hypercube around h restricted by radius}

\item{p.df}{Dataframe which describes the parameters of the prior distributions for each HLA}
}
\description{
Performs one time step in the MH chain:
}

