% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/immunogen_predictive_prior.r
\name{plot_beta}
\alias{plot_beta}
\title{Plotting a beta distribution with a given mode and standard deviation}
\usage{
plot_beta(M, SD)
}
\arguments{
\item{M}{mode}

\item{SD}{standard deviation}
}
\description{
Plotting a beta distribution with a given mode and standard deviation
}

