% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/immunogen_plots.r
\name{plot_posteriors}
\alias{plot_posteriors}
\title{Get a plot of the marginal posteriors for each pHLA combination (all HLAs, one peptide p)}
\usage{
plot_posteriors(chain, nCol = 3, fileName)
}
\arguments{
\item{chain}{The MH chain in question}

\item{nCol}{The number of columns in the output figure}

\item{fileName}{Filename for the figure}
}
\description{
Implementation of the Metropolis-Hastings algorithm to find posterior distributions of pHLA immunogenicity for a single peptide
}

